cloud_function/requirements.txt
Flask==2.0.1
python-dotenv
functions-framework
gunicorn
